requirements.txt
pytest
requests
allure-pytest==2.13.5
allure-python-commons==2.13.5